Refactor: getWeeklyEventSlots 함수 utils로 빼내고 dailyEventsMap의 필터 로직을 getWeeklyEventSlots에 넣어 성능 최적화
